+++ pkg/action/start.go
@@ -6,11 +6,13 @@ import (
 
 	"github.com/urfave/cli"
 
+	"github.com/andrskom/jwa-console/pkg/tag"
 	"github.com/andrskom/jwa-console/pkg/timeline"
 )
 
 func Start(
 	timelineComponent *timeline.Component,
+	tagComponent *tag.Component,
 ) func(c *cli.Context) error {
 	return func(c *cli.Context) error {
 		taskID := c.Args().Get(0)
@@ -32,10 +34,19 @@ func Start(
 			opts.UsePrevDescription = true
 		}
 
-		model, err := timelineComponent.Start(taskID, opts)
+		model, err := timelineComponent.BuildModel(taskID, opts)
 		if err != nil {
 			return err
 		}
+		if err := tagComponent.SetTag(c.String("t"), c.Bool("nt"), model); err != nil {
+			return err
+		}
+
+		model, err = timelineComponent.Start(model)
+		if err != nil {
+			return err
+		}
+
 		fmt.Printf(`Start task %s %s
 `, model.Issue.Key, model.Issue.Fields.Summary)
 		return nil