+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-beta] - 2019-07-07
+### Added
+- Aliases for task's commands.
+- Configuration command.
+- Change task command.
+- Tags for work.
+- In progress statuses check before start.
+
+## [0.0.4-alpha] - 2019-06-13
+### Added
+- Edit of task.
+
+## [0.0.3-alpha] - 2019-06-13
+### Fixed
+- Time format.
+
 ## [0.0.2-alpha] - 2019-06-12
 ### Added
 - Wide info in show.